Add two options to resume: the ability to specify a run-id and the ability to only clone tasks
The scenario for this use is mainly for external schedulers that are
capable of "resuming" failed runs by only re-executing the ones that failed
or didn't execute the first time. We need a way for Metaflow to "clone" the
tasks that are not re-executed.

Also took the opportunity to fix a few TODOs related to status updates for local runs
as well as some tiny cleanups.